One controller for one fxml, changing background
Implemented one controller for one fxml-file, so that one controller does not have to work with all the fxml-files. This was causing problems because of loading objects that were null within a new window but were not necessary. Also added the possibility to change the backgroundcolor. The current colors are just for testing.