Started better Exception handling
Tried to catch the exception when it is thrown. The handling is right now only to e.printStacktrace() and might be changed in the future. Also changed the code in a way that there will be less errors because of conditions that has to be true otherwise following steps will not be executed.